[REFACTOR] course api (#25)
* [RED] refactor tests to route course

* [GREEN] view course api moved to CourseController

* [REFACTOR] fix code smell

* [REFACTOR] optimize query

* [REFACTOR] fix code smell attempt 1

* [REFACTOR] improve api url

* [REFACTOR] fix var naming
